clean up project
