+++ Example/Images.xcassets/AppIcon.appiconset/Contents.json
