--- Example/Podfile
@@ -1,11 +1,0 @@
-use_frameworks!
-
-target 'TactileSlider_Example' do
-  pod 'TactileSlider', :path => '../'
-
-  target 'TactileSlider_Tests' do
-    inherit! :search_paths
-
-    
-  end
-end
