--- Example/TactileSlider.xcworkspace/contents.xcworkspacedata
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "group:../Package.swift">
-   </FileRef>
-   <FileRef
-      location = "group:TactileSlider.xcodeproj">
-   </FileRef>
-   <FileRef
-      location = "group:Pods/Pods.xcodeproj">
-   </FileRef>
-</Workspace>
