+++ TactileSlider.podspec
@@ -8,21 +8,13 @@
 
 Pod::Spec.new do |s|
   s.name             = 'TactileSlider'
-  s.version          = '2.0.0'
+  s.version          = '2.0.1'
   s.summary          = 'Easy-to-grab slider control inspired by Control Center and HomeKit.'
-
-# This description is used to generate tags and improve search results.
-#   * Think: What does it do? Why did you write it? What is the focus?
-#   * Try to keep it short, snappy and to the point.
-#   * Write the description between the DESC delimiters below.
-#   * Finally, don't worry about the indent, CocoaPods strips it!
-
   s.description      = <<-DESC
 A slider control designed to be easy to grab and use because it can be dragged or tapped from anywhere along its track, similar to the sliders in Control Center and HomeKit.
                        DESC
 
   s.homepage         = 'https://github.com/daprice/iOS-Tactile-Slider'
-  # s.screenshots     = 'www.example.com/screenshots_1', 'www.example.com/screenshots_2'
   s.license          = { :type => 'MIT', :file => 'LICENSE' }
   s.author           = { 'Dale Price' => 'daprice@mac.com' }
   s.source           = { :git => 'https://github.com/daprice/iOS-Tactile-Slider.git', :tag => s.version.to_s }
@@ -31,13 +23,6 @@ A slider control designed to be easy to grab and use because it can be dragged o
   s.ios.deployment_target = '8.0'
   s.swift_version = '5.0'
 
-  s.source_files = 'TactileSlider/Classes/**/*'
-  
-  # s.resource_bundles = {
-  #   'TactileSlider' => ['TactileSlider/Assets/*.png']
-  # }
-
-  # s.public_header_files = 'Pod/Classes/**/*.h'
+  s.source_files = 'TactileSlider/*'
   s.frameworks = 'UIKit'
-  # s.dependency 'AFNetworking', '~> 2.3'
 end
