+++ TactileSlider.xcodeproj/project.xcworkspace/contents.xcworkspacedata
