+++ TactileSlider.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
