+++ TactileSlider/TactileSliderLayerRenderer.swift
